added casts to resolve 64-bit build issues
git-svn-id: http://svn.osafoundation.org/pyicu/trunk@84 17bc4a61-6af7-0310-99d3-950ecb8e7402

+++ common.cpp
@@ -506,7 +506,7 @@ Formattable *toFormattable(PyObject *arg)
         return new Formattable(i);
 
     if (!parseArg(arg, "L", &l))
-        return new Formattable(l);
+      return new Formattable((int64_t) l);
 
     if (!parseArg(arg, "c", &s))
         return new Formattable(s);
