+++ numberformat.cpp
@@ -497,7 +497,7 @@ static PyObject *t_numberformat_format(t_numberformat *self, PyObject *args)
         }
         if (!parseArgs(args, "L", &l))
         {
-            self->object->format(l, _u);
+	  self->object->format((int64_t) l, _u);
             return PyUnicode_FromUnicodeString(&_u);
         }
         break;
@@ -517,7 +517,7 @@ static PyObject *t_numberformat_format(t_numberformat *self, PyObject *args)
         if (!parseArgs(args, "LP", TYPE_CLASSID(FieldPosition),
                        &l, &fp))
         {
-            self->object->format(l, _u, *fp);
+	  self->object->format((int64_t) l, _u, *fp);
             return PyUnicode_FromUnicodeString(&_u);
         }
         break;
@@ -537,7 +537,7 @@ static PyObject *t_numberformat_format(t_numberformat *self, PyObject *args)
         if (!parseArgs(args, "LUP", TYPE_CLASSID(FieldPosition),
                        &l, &u, &fp))
         {
-            self->object->format(l, *u, *fp);
+	  self->object->format((int64_t) l, *u, *fp);
             Py_RETURN_ARG(args, 1);
         }
         break;
@@ -1672,7 +1672,7 @@ static PyObject *t_rulebasednumberformat_format(t_rulebasednumberformat *self,
         if (!parseArgs(args, "LSP", TYPE_CLASSID(FieldPosition),
                        &l, &u, &_u, &fp))
         {
-            STATUS_CALL(self->object->format(l, *u, _v, *fp, status));
+	  STATUS_CALL(self->object->format((int64_t) l, *u, _v, *fp, status));
             return PyUnicode_FromUnicodeString(&_v);
         }
       case 4:
@@ -1691,7 +1691,7 @@ static PyObject *t_rulebasednumberformat_format(t_rulebasednumberformat *self,
         if (!parseArgs(args, "LSUP", TYPE_CLASSID(FieldPosition),
                        &l, &u, &_u, &v, &fp))
         {
-            STATUS_CALL(self->object->format(l, *u, *v, *fp, status));
+	  STATUS_CALL(self->object->format((int64_t) l, *u, *v, *fp, status));
             Py_RETURN_ARG(args, 2);
         }
         break;
